+++ test/integration/infra/scale-testing/deployments/providers.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     helm = {
       source  = "hashicorp/helm"
-      version = "2.13.1"
+      version = "2.16.1"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"